requirements.txt
llama-index==0.12.52
llama-index-core==0.12.52.post1
openai
llama-parse==0.6.65
cohere
pandas
ragas
deepeval
llama-index-llms-groq==0.3.2
llama-index-embeddings-huggingface==0.5.5
langchain-huggingface
langchain-openai
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai-like==0.4.0
plotly
kaleido
streamlit
llama-index-llms-openrouter
python-dotenv
